requirements.txt
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
dateparser==0.7.2
decorator==4.4.1
defusedxml==0.6.0
en-core-web-sm==2.2.5
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
glob2==0.7
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-pubsub==1.2.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
idna==2.8
import-ipynb==0.1.3
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
jieba3k==0.35.1
Jinja2==2.10.3
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
newspaper3k==0.2.8
nltk==3.4.5
notebook==6.0.2
numba==0.48.0
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plotly==4.5.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pyodbc==4.0.28
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.0
qtconsole==4.6.0
regex==2020.1.8
requests==2.22.0
requests-file==1.4.3
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
spacy==2.2.4
spacy-lookups-data==0.2.0
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.0
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
tzlocal==2.0.0
umap-learn==0.3.10
urllib3==1.25.7
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
zipp==0.6.0
